+++ packages/wiz-ui-react/src/components/base/inputs/password/components/password-input.tsx
@@ -23,7 +23,7 @@ type Props = BaseProps & {
   width?: string;
   autocomplete?: Extract<AutoCompleteKeys, "currentPassword" | "newPassword">;
   error?: boolean;
-  onChange: (value: string) => void;
+  onChange?: (value: string) => void;
 } & Omit<PrivateBaseInputProps, "onChange">;
 
 const PasswordInput = forwardRef<HTMLInputElement, Props>(
@@ -67,10 +67,10 @@ const PasswordInput = forwardRef<HTMLInputElement, Props>(
           expand={expand}
           width={width}
           error={isError}
-          type={isPasswordVisible ? "text" : "password"}
+          type={!disabled && isPasswordVisible ? "text" : "password"}
           autoComplete={autocomplete}
-          space-type="right"
-          onChange={(e) => onChange(e.target.value)}
+          spaceType="right"
+          onChange={(e) => onChange?.(e.target.value)}
           {...props}
         />
         <button
@@ -78,15 +78,15 @@ const PasswordInput = forwardRef<HTMLInputElement, Props>(
           aria-label={ARIA_LABELS.PASSWORD_VISIBLE_TOGGLE}
           className={styles.passwordVisibleIconStyle}
           onClick={() => setIsPasswordVisible(!isPasswordVisible)}
+          disabled={disabled}
         >
-          <div
+          <WizIEye
             className={clsx(
-              isPasswordVisible && styles.passwordVisibleIconActiveStyle
+              !disabled &&
+                isPasswordVisible &&
+                styles.passwordVisibleIconActiveStyle
             )}
-            style={{ display: "flex" }}
-          >
-            <WizIEye />
-          </div>
+          />
         </button>
       </div>
     );