+++ packages/wiz-ui-react/src/components/base/text-area/components/text-area.tsx
@@ -24,7 +24,7 @@ type Props = BaseProps & {
   minWidth?: string;
   maxHeight?: string;
   minHeight?: string;
-  onChange: (value: string) => void;
+  onChange?: (value: string) => void;
 };
 
 const TextArea = forwardRef<HTMLTextAreaElement, Props>(
@@ -71,11 +71,10 @@ const TextArea = forwardRef<HTMLTextAreaElement, Props>(
         className={clsx(
           className,
           styles.textAreaStyle,
-          styles.textAreaVariantStyle[disabled ? "disabled" : "default"],
           expand && styles.textAreaExpandStyle,
           inputBorderStyle[getInputBorderStyleKey(isError)]
         )}
-        onChange={(e) => onChange(e.target.value)}
+        onChange={(e) => onChange?.(e.target.value)}
       />
     );
   }